Just missing auth

+++ apis/organisationOverviewAPI.ts
@@ -12,7 +12,15 @@ export const fetchAllOrganisationsRequest = async (userId: string) => {
   return res.json();
 };
 
-export const deleteOrganisationRequest = async (organisationId: number) => {
+export const deleteOrganisationRequest = async (userId: string | null, organisationId: number) => {
+  if (userId === null) {
+    throw new Error("FATAL FEJL: Bruger-ID er ikke korrekt initialiseret i din session.");
+  }
+  //Do so only the owner can delete the organisation Implement with auth
+  if (userId !== null) {
+    throw new Error("Fejl: Kun ejeren af organisationen kan slette den");
+  }
+
   const url = `${BASE_URL}/organizations/${organisationId}`;
   const res = await fetch(url, {
     method: "DELETE",
